+++ packages/myst-frontmatter/src/page/validators.ts
@@ -7,10 +7,12 @@ import {
   validateBoolean,
   validateObject,
   validateUrl,
+  validationWarning,
 } from 'simple-validators';
 import { validateProjectAndPageFrontmatterKeys } from '../project/validators.js';
 import { PAGE_FRONTMATTER_KEYS, type PageFrontmatter } from './types.js';
 import { validateKernelSpec } from '../kernelspec/validators.js';
+import { validateExecute } from '../execute/validators.js';
 import { validateJupytext } from '../jupytext/validators.js';
 import { FRONTMATTER_ALIASES } from '../site/types.js';
 
@@ -52,11 +54,23 @@ export function validatePageFrontmatterKeys(value: Record<string, any>, opts: Va
   if (defined(value.jupytext)) {
     output.jupytext = validateJupytext(value.jupytext, incrementOptions('jupytext', opts));
   }
+  if (defined(value.execute)) {
+    output.execute = validateExecute(value.execute, incrementOptions('execute', opts));
+  }
   if (defined(value.skip_execution)) {
-    output.skip_execution = validateBoolean(
-      value.skip_execution,
-      incrementOptions('skip_execution', opts),
-    );
+    output.execute ??= {};
+    if (defined(output.execute.skip)) {
+      validationWarning(
+        `both execute.skip and deprecated skip_execution are defined, taking execute.skip`,
+        opts,
+      );
+    } else {
+      validationWarning(`skip_execution is deprecated in favour of execute.skip`, opts);
+      output.execute.skip = validateBoolean(
+        value.skip_execution,
+        incrementOptions('skip_execution', opts),
+      );
+    }
   }
   if (defined(value.enumerator)) {
     output.enumerator = validateString(value.enumerator, incrementOptions('enumerator', opts));