+++ packages/myst-directives/CHANGELOG.md
@@ -1,6 +1,21 @@
 # myst-directives
 
+## 1.6.1
+
+### Patch Changes
+
+- 1735db4: Add support for iframe title.
+- Updated dependencies [1735db4]
+- Updated dependencies [30acc57]
+  - myst-spec-ext@1.9.0
+  - myst-common@1.9.0
+
 ## 1.6.0
+
+### Patch Changes
+
+- myst-common@1.8.4
+- myst-spec-ext@1.8.4
 
 ## 1.5.17