+++ packages/myst-roles/CHANGELOG.md
@@ -1,6 +1,22 @@
 # myst-roles
 
+## 1.6.1
+
+### Patch Changes
+
+- 6fe5960: Set the label for terms if not modified
+- Updated dependencies [1735db4]
+- Updated dependencies [30acc57]
+  - myst-spec-ext@1.9.0
+  - myst-common@1.9.0
+
 ## 1.6.0
+
+### Patch Changes
+
+- e293e72: Add node information to doc-role warning
+  - myst-common@1.8.4
+  - myst-spec-ext@1.8.4
 
 ## 1.5.17