+++ packages/myst-to-html/CHANGELOG.md
@@ -1,10 +1,18 @@
 # myst-to-html
 
+## 1.6.1
+
+### Patch Changes
+
+- Updated dependencies [30acc57]
+  - myst-common@1.9.0
+
 ## 1.6.0
 
 ### Patch Changes
 
 - cd8ee79: Update MarkdownIt to 13 for all dependencies, add types to dependencieswq
+  - myst-common@1.8.4
 
 ## 1.5.17